Update mediate_timeout to use float type instead of int.

+++ src/browserist/helper/timeout.py
@@ -9,7 +9,7 @@ def should_continue(settings: BrowserSettings) -> bool:
     ])
 
 
-def mediate_timeout(settings: BrowserSettings, timeout: int | None) -> int:
+def mediate_timeout(settings: BrowserSettings, timeout: float | None) -> float:
     return settings.timeout.seconds if timeout is None else timeout
